classes/sdcard_image-rpi: rework to not use loop mounts
Avoid the use of loop mounts so you don't have to do any setup on the
machine before being able to create an SD card image. Instead of loop
mounting the boot filesystem, populate it with mtools and then use dd to
copy it into the SD card image at the appropriate offsets. The root
filesystem is handled similarly.

Signed-off-by: Paul Eggleton <paul.eggleton@linux.intel.com>