classes/sdcard_image-rpi: change to image type class
Instead of being inherited by a specific image recipe, this class can
now be added to the IMAGE_CLASSES variable (together with image_types)
allowing it to be used with any image. This means we can remove
"inherit image", and we can also remove the IMAGE_FSTYPES modification
since this is not meant to be done by this kind of class, and in any
case is already handled by the raspberrypi machine configuration
anyway.

Signed-off-by: Paul Eggleton <paul.eggleton@linux.intel.com>